requirements.txt
gitpython>=3.1.40
rich>=13.7.0
prompt_toolkit>=3.0.43
click>=8.1.7
pygments>=2.17.2 